LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-autofocus-expected.txt

      
 


FAIL autofocus when a modal dialog is opened assert_equals: expected Element node <button id="outer-button" autofocus=""></button> but got Element node <button id="another-button" autofocus=""></button>
